// src/core/api/generated/spring/schemas/AvailableTimeSlotDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type AvailableTimeSlotDTOStatus = typeof AvailableTimeSlotDTOStatus[keyof typeof AvailableTimeSlotDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AvailableTimeSlotDTOStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;